+++ .eslintrc.json
@@ -3,7 +3,7 @@
   "rules": {
     "max-len": ["warn", { "code": 120 }],
     "linebreak-style": "off",
-    "react/destructuring-assignment": [false],
+    "react/destructuring-assignment": [0],
     "react/prefer-stateless-function": "off",
     "react/prop-types": "off",
     "no-unused-vars": ["warn", { "args": "none" }],
